\name{FigB3}
\alias{FigB3}
\title{A function to plot Figure B.3}
\description{This function plots the age-adjusted trends in overweight prevalence in women amidst seasonal and sampling background fluctuations by occupation groups shown in Figure B.3. The fluctuations and trend line for high occupations are in red, and in purple and blue for medium and low occupations, respectively. The trends are predicted using quarterly data (4 prevalence measurements per year) over 12 years. The plot is automatically generated by calling this function and is outputted to a .png file called "plot1f.png".
    }
\usage{FigB3()}

\author{Yun(Albee) Ling, Sierra Costanza, Bingjing Tang <yling1@jhu.edu, scostan1@jhu.edu, btang8@jhu.edu>}

\examples{
data(Xf)
FigB3()
}