\name{FigC1}
\alias{FigC1}
\title{A function to plot Figure C.1}
\description{This function plots a histogram of the currently-smoked pack-years among medium occupation women (from \code{Xf}), shown in Figure C.1 in the Model Selection section of the final report. The output is a .png file called "dhist.png".
    }
\usage{FigC1()}

\author{Yun(Albee) Ling, Sierra Costanza, Bingjing Tang <yling1@jhu.edu, scostan1@jhu.edu, btang8@jhu.edu>}

\examples{
data(Xf)
FigC1()
}