\name{finalrpack}
\alias{finalrpack-package}
\docType{package}
\title{How Much Ice Do You Need?}
\description{Includes function and data deliverables for our sponsor }
\details{ 
shc: Returns dilution and temperature of resulting drink based based on volume of ice and soda, as well as specific heat capacity of soda
problem2: datasets that includes medal counts for Winter Olympics for different countries
}

\author{Joyce Tan jtan21@jhu.edu
Yen Theng Tan yen@jhu.edu}

\keyword{package}

\examples{}

