\name{obesity_predict}
\alias{obpredict}
\docType{package}
\title{
A package with obesity-related data and a function to predict obesity rates
}
\description{
The included data contains data on 14 independent variables for 190 countries. The included function will predict obesity rate based on an optimally chosen regression model, and inputted independent variable values.
}
\details{
\tabular{ll}{
Package: \tab obesity_prediction\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-12-19\cr
License: \tab None\cr
}
Load the data frame "obesity_update" and use it to explore obesity data on your own. Use the predict_obesity function to experiment with values of independent variables.
}
\author{
Shannon Cebron, Michael Weinberger, Zhendan Zhu

Who to complain to <scebron@cis.jhu.edu>
}
\references{
CIA World Factbook
WHO Health Database
}
\keyword{ package }
