\name{predicted_yield}
\alias{predicted_yield}
\title{predicted_yield
}
\description{
This dataset is the predicted yeild curve we obtained by plugging the real data into yield polynomial regression model.
}
\usage{
data(predicted_yield)
}

\arguments{
  \item{predicted_yield}{
This is the predicted yeilds we obtained by plugging the real data into yield polynomial regression model.
}
 
}
\references{
Please refer to Blackrock's final report
}
\author{
Shihong Li
}

\examples{

data(predicted_yield)
}
\keyword{datasets}