\name{sprite}
\alias{sprite}
\docType{data}
\title{
Experimental results for Sprite
}
\description{
Collated preferences of subjects for Sprite, at different time parameters and different ice proportions}
\usage{data(coke)}
\format{

The numbers in the cell represent the total points that the participants gave (for each participant - 3 being the most preferred, 2 the second most, and 1 the least), for a given time parameter (specified by the left column), and for a given proportion of ice (top row)
}
\details{ }
\source{
Team McDonald's}
\references{
}
\examples{
data(sprite)

}
\keyword{datasets}
