\name{finalrpack}
\alias{finalrpack-package}
\docType{package}
\title{How Much Ice Do You Need?}
\description{Includes function and data deliverables for our sponsor }
\details{ 
\itemize
\item shc: Returns dilution and temperature of resulting drink based based on volume of ice and soda, as well as specific heat capacity of soda
\item problem2: datasets that includes medal counts for Winter Olympics for different countries
}

\author{
\itemize
\item Joyce Tan jtan21@jhu.edu
\item Yen Theng Tan yen@jhu.edu}

\keyword{package}

\examples{}

